# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EmbodiedCloud API",
    "description": "Browser-first Isaac Lab workspace control plane",
    "version": "0.4.0"
  },
  "paths": {
...
    "/api/workspaces": {
      "get": {
        "tags": [
          "workspaces"
        ],
        "summary": "List Workspaces",
        "operationId": "list_workspaces_api_workspaces_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkspaceOut"
                  },
                  "type": "array",
                  "title": "Response List Workspaces Api Workspaces Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "workspaces"
        ],
        "summary": "Create Workspace",
        "operationId": "create_workspace_api_workspaces_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{workspace_id}": {
      "get": {
        "tags": [
          "workspaces"
        ],
        "summary": "Get Workspace",
        "operationId": "get_workspace_api_workspaces__workspace_id__get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "workspaces"
        ],
        "summary": "Delete Workspace",
        "description": "\u00a712\uff1adurable DESTROY \u2014\u2014 enqueue DESTROY operation\uff1bworker \u6267\u884c\n\uff08\u5e42\u7b49 tombstone\uff1b\u91cd\u542f\u4e0d\u4e22\u6e05\u7406\uff09\u3002",
        "operationId": "delete_workspace_api_workspaces__workspace_id__delete",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{workspace_id}/access": {
      "get": {
        "tags": [
          "workspaces"
        ],
        "summary": "Get Workspace Access",
        "operationId": "get_workspace_access_api_workspaces__workspace_id__access_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceAccessOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{workspace_id}/start": {
      "post": {
        "tags": [
          "workspaces"
        ],
        "summary": "Start Workspace",
        "operationId": "start_workspace_api_workspaces__workspace_id__start_post",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{workspace_id}/stop": {
      "post": {
        "tags": [
          "workspaces"
        ],
        "summary": "Stop Workspace",
        "description": "\u00a712\uff1adurable STOP \u2014\u2014 enqueue STOP operation\uff1bworker \u6267\u884c\uff08fast-path \u540c\u6b65\n\u5904\u7406\uff0c\u4efb\u52a1\u6301\u4e45\u5316\uff1a\u63a7\u5236\u9762\u91cd\u542f\u540e cleanup \u4e0d\u4e22\u5931\uff09\u3002",
        "operationId": "stop_workspace_api_workspaces__workspace_id__stop_post",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{workspace_id}/logs": {
      "get": {
        "tags": [
          "workspaces"
        ],
        "summary": "Workspace Logs",
        "description": "runtime \u65e5\u5fd7\uff08owner \u9694\u79bb\uff1bprovider.logs\uff0c\u65e0 runtime \u65f6\u8fd4\u56de\u7a7a\uff09\u3002",
        "operationId": "workspace_logs_api_workspaces__workspace_id__logs_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "tail",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 200,
              "title": "Tail"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Workspace Logs Api Workspaces  Workspace Id  Logs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{workspace_id}/demo-checkpoint": {
      "post": {
        "tags": [
          "workspaces"
        ],
        "summary": "Create Demo Checkpoint",
        "description": "mock \u6f14\u793a\u6a21\u5f0f\u4e13\u7528\uff1a\u5728 workspace \u76ee\u5f55\u751f\u6210\u4e00\u4efd\u6f14\u793a checkpoint\u3002\n\n\u771f\u5b9e\u8bad\u7ec3\u4ea7\u51fa\u9700\u8981 GPU runtime\uff1b\u6f14\u793a\u73af\u5883\u7528\u8be5\u7aef\u70b9\u6a21\u62df\u300c\u8bad\u7ec3\u5b8c\u6210\u300d\uff0c\n\u8ba9 Sim2Real \u90e8\u7f72\u6d41\u7a0b\uff08artifact \u2192 checksum \u2192 deploy \u2192 edge \u6821\u9a8c\uff09\u53ef\u7aef\u5230\u7aef\u8d70\u901a\u3002\n\u975e mock provider \u62d2\u7edd\uff08\u771f\u5b9e\u73af\u5883\u5fc5\u987b\u7528\u771f\u5b9e\u4ea7\u51fa\uff0c\u4e0d\u5f97\u4f2a\u9020\uff09\u3002",
        "operationId": "create_demo_checkpoint_api_workspaces__workspace_id__demo_checkpoint_post",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoCheckpointOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/streaming/workspace/{workspace_id}": {
      "get": {
        "tags": [
          "streaming"
        ],
        "summary": "List Streaming Sessions",
        "operationId": "list_streaming_sessions_api_streaming_workspace__workspace_id__get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StreamingOut"
                  },
                  "title": "Response List Streaming Sessions Api Streaming Workspace  Workspace Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DemoCheckpointOut": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          }
        },
        "type": "object",
        "required": [
          "path",
          "size_bytes",
          "sha256"
        ],
        "title": "DemoCheckpointOut",
        "description": "mock \u6f14\u793a\u6a21\u5f0f\u7684\u6a21\u62df\u8bad\u7ec3\u4ea7\u51fa\uff1a\u4f9b Sim2Real \u90e8\u7f72\u6d41\u7a0b\u7aef\u5230\u7aef\u6f14\u793a\u3002"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StreamingOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "signal_port": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signal Port"
          },
          "media_port": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Media Port"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "workspace_id",
          "status",
          "signal_port",
          "media_port",
          "error_message",
          "created_at",
          "updated_at"
        ],
        "title": "StreamingOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkspaceAccessOut": {
        "properties": {
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "ide_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ide Url"
          },
          "ide_password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ide Password"
          },
          "stream_hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stream Hint"
          },
          "signal_port": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signal Port"
          },
          "media_port": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Media Port"
          }
        },
        "type": "object",
        "required": [
          "workspace_id",
          "status",
          "ide_url",
          "ide_password",
          "stream_hint",
          "signal_port",
          "media_port"
        ],
        "title": "WorkspaceAccessOut"
      },
      "WorkspaceCreate": {
        "properties": {
          "template_id": {
            "type": "string",
            "title": "Template Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "auto_start": {
            "type": "boolean",
            "title": "Auto Start",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "template_id"
        ],
        "title": "WorkspaceCreate"
      },
      "WorkspaceOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "template_id": {
            "type": "string",
            "title": "Template Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "organization_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "gpu_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gpu Index"
          },
          "gpu_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gpu Name"
          },
          "gpu_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gpu Id"
          },
          "ide_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ide Url"
          },
          "stream_hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stream Hint"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "stopped_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stopped At"
          },
          "accumulated_seconds": {
            "type": "integer",
            "title": "Accumulated Seconds"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "template_id",
          "user_id",
          "organization_id",
          "status",
          "provider",
          "gpu_index",
          "gpu_name",
          "gpu_id",
          "ide_url",
          "stream_hint",
          "error_message",
          "created_at",
          "started_at",
          "stopped_at",
          "accumulated_seconds"
        ],
        "title": "WorkspaceOut"
      }
    }
  }
}